backend/requirements/base.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0
pyotp>=2.9,<3.0
qrcode[pil]>=7.4,<8.0

# Database
psycopg[binary]>=3.1,<4.0

# Environment & Config
python-decouple>=3.8,<4.0
dj-database-url>=2.1,<3.0

# File handling
Pillow>=10.0,<11.0

# PDF Generation
WeasyPrint>=60.0,<61.0

# Email
django-anymail[mailgun,sendgrid]>=10.0,<11.0

# Utils
django-filter>=23.0,<24.0
python-dateutil>=2.8,<3.0

# API Documentation
drf-spectacular>=0.27,<1.0
